ticket 47352 - Abandon old password changes prior to syncing
We currently go through abandoning old password changes in the
PassSync service right after we process sending changes to DS.  In
some cases, this can cause us to send old password changes to DS
that have already passed the "Time To Live" setting.

This patch moved the abandoning of old password to happen right
before we send password to DS.  This will ensure that we don't
send old password changes that may be queued up.

https://fedorahosted.org/389/ticket/47352

Reviewed by Noriko.