Ticket #363 - Passsync/Winsync handles passwords with 8-th bit characters incorrectly
Bug description: Passhook plugin used to store the password in
the intermediate file passhook.dat using _snprintf which just
converts wchar in ascii to char without considering the non-
ascii characters.

Fix description: Instead of using _snprintf, WideCharToMultiByte
is called to convert the Microsoft internal character set to
UTF-8, which is valid in LDAP.

Also, in SyncPasswords (PassSync), it adds LDAP_INAPPROPRIATE_
AUTH to the condition to send the password change on Windows to
the Directory server.  Bind returns LDAP_INAPPROPRIATE_AUTH,
when a password is not in the entry for SIMPLE auth.  PassSync
should be able to send the password for the case, as well.

https://fedorahosted.org/389/ticket/363

Reviewed by Rich (Thanks!!)